Added VecZnxBig<FFT64> ops
